resolved comments, cleanup
